This should work well. Time to install the solar arrays.
Added some javacript sample programs. Very usefull to test things with
node in commandline.
This version loads a configuration json with each miner information, and
stores it in global context. Then at runtime, it updates it from the
Hiveos servers and then starts or stops mining depending on the current
available PV power.